Supported set high definition monitor log status for eip.

+++ vpc/include/alibabacloud/vpc/model/DescribeEipMonitorDataResult.h
@@ -36,7 +36,7 @@ namespace AlibabaCloud
 				{
 					int eipPackets;
 					int eipBandwidth;
-					int eipFlow;
+					long eipFlow;
 					long eipTX;
 					std::string timeStamp;
 					long eipRX;
